Click to zoom in, improved colouring
Store plot location and zoom in the scene object.
Sprite generation is now a proper method of the scene object.
Determine the max/min iterations for the plot and scale colouring
accordinly so the plot always uses the full range of colours.
Click handler calculates where was clicked relative to the plot and
generates a new plot centered on that and at a higher zoom level.